fix : 05_abstract class 수정
