+++ lib/04_inheitance/library.dart
@@ -25,56 +25,3 @@ class Book {
   }
 }
 
-class Fiction extends Book {
-  String subgenre; //하위 장르
-  String creativity; //창의성
-  String imagination; //상상력
-
-  Fiction(
-    super.title,
-    super.author,
-    super.genre,
-    super.bookId,
-    this.subgenre,
-    this.creativity,
-    this.imagination,
-  );
-
-  @override
-  void openBook() {
-    print('책을 뜯었습니다.');
-  }
-
-  @override
-  void checkRefund() {
-    print('오픈시 환불불가');
-  }
-}
-
-class NonFiction extends Book {
-  String subject; //주제
-  String reality; // 실제
-  String infoTransfer; //정보전달
-  String eduValue; // 교육가치
-
-  NonFiction(
-    super.title,
-    super.author,
-    super.genre,
-    super.bookId,
-    this.subject,
-    this.reality,
-    this.infoTransfer,
-    this.eduValue,
-  );
-
-  @override
-  void openBook() {
-    print('책을 뜯었습니다.');
-  }
-
-  @override
-  void checkRefund() {
-    print('오픈시 환불불가');
-  }
-}
