--- lib/05_abstract_class_interface/exam/book.dart
@@ -1,15 +1,0 @@
-class Book {
-  String name;
-  int price;
-  String color;
-  String isbn;
-
-  Book ({
-    required this.name,
-    required this.price,
-    required this.color,
-    required this.isbn,
-
-  });
-
-}
