--- lib/05_abstract_class_interface/exam/computer.dart
@@ -1,15 +1,0 @@
-class Computer {
-  String name;
-  int price;
-  String color;
-  String isbn;
-
-  Computer ({
-    required this.name,
-    required this.price,
-    required this.color,
-    required this.isbn,
-
-  });
-
-}
