--- lib/05_abstract_class_interface/exam/tangible_asset.dart
@@ -1,3 +1,0 @@
-abstract class TangibleAsset {
-
-}
