--- lib/05_abstract_class_interface/exam/thing.dart
@@ -1,4 +1,0 @@
-abstract interface class Thing {
-  double get weight;
-  set weight (double value);
-}
