+++ lib/07_instance/copy.dart
@@ -33,3 +33,12 @@ class Person {
 //     age: age ?? this.age,
 //   );
 // }
+
+bool operator(Object other) {
+  return identical(this, other) ||
+    other is Person && runtimeType == other.runtimeType
+      && name == other.name
+      && hp == other.hp;
+
+
+}
